+++ lib/src/generators/world/opinions/base_opinion_generator.dart
@@ -1,0 +1,45 @@
+import '../../base/list_item_generator.dart';
+
+/// A class that generates generic opinions about races
+class BaseOpinionGenerator extends ListItemGenerator<String> {
+  static const _opinions = [
+    'lazy gluttons',
+    'resourceful fighters',
+    'reclusive and suspicious',
+    'very hospitable',
+    'cheating liars',
+    'unassuming and reliable',
+    'devoted and religious',
+    'religious zealots',
+    'clever and resourceful',
+    'dangerous and ambitious',
+    'warmongers',
+    'practical survivalists',
+    'skeptical and paranoid',
+    'strange and migratory',
+    'cursed beings',
+    'always cheerful',
+    'bloodthirsty hunters',
+    'cruel tricksters',
+    'interesting with rich culture',
+    'honorable fighters',
+    'uncaring and cruel',
+    'expert artisans',
+    'dependable traders',
+    'terrifying fighters',
+    'untrustworthy mercenaries',
+    'expert tacticians',
+    'greedy hoarders',
+    'xenophobic zealots',
+    'self-destructively cautious',
+    'strong and fearless',
+    'natural adventurers',
+    'thieving scoundrels',
+    'stubbornly determined',
+    'terrifying',
+    'the greatest warriors',
+    'proud and honorable'
+  ];
+
+  BaseOpinionGenerator() : super(_opinions);
+}