+++ lib/src/subtypes/races/dwarf/dwarf.dart
@@ -14,8 +14,10 @@ import '../../../generators/npcs/physical/face/base_face_generator.dart';
 import '../../../generators/npcs/physical/hair/base_hair_generator.dart';
 import '../../../generators/npcs/physical/skin/base_skin_generator.dart';
 import '../../../generators/npcs/physical/special_features/base_special_feature_generator.dart';
+import '../../../generators/world/opinions/base_opinion_generator.dart';
 import '../race.dart';
 import 'names/dwarf_name_generator.dart';
+import 'opinions/dwarf_opinion_generator.dart';
 import 'quirks/dwarf_quirks_generator.dart';
 
 /// A class that represents the dwarf race
@@ -81,4 +83,10 @@ class Dwarf implements Race {
   @override
   IGenerator<String> getSpecialFeatureGenerator(Gender gender) =>
       BaseSpecialFeatureGenerator();
+
+  @override
+  IGenerator<String> getOpinionGenerator() => MultipleGenerator([
+        BaseOpinionGenerator(),
+        DwarfOpinionGenerator(),
+      ]);
 }