+++ lib/src/subtypes/races/elf/elf.dart
@@ -11,10 +11,12 @@ import '../../../generators/npcs/physical/build/light_build_generator.dart';
 import '../../../generators/npcs/physical/eyes/base_eyes_generator.dart';
 import '../../../generators/npcs/physical/face/good_looking_face_generator.dart';
 import '../../../generators/npcs/physical/special_features/base_special_feature_generator.dart';
+import '../../../generators/world/opinions/base_opinion_generator.dart';
 import '../race.dart';
 import 'alignment/elven_alignment_generator.dart';
 import 'hair/elven_hair_generator.dart';
 import 'names/elf_name_generator.dart';
+import 'opinions/elf_opinion_generator.dart';
 import 'quirks/elf_quirks_generator.dart';
 import 'skin/elf_skin_generator.dart';
 
@@ -82,4 +84,10 @@ class Elf implements Race {
   @override
   IGenerator<String> getSpecialFeatureGenerator(Gender gender) =>
       BaseSpecialFeatureGenerator();
+
+  @override
+  IGenerator<String> getOpinionGenerator() => MultipleGenerator([
+        BaseOpinionGenerator(),
+        ElfOpinionGenerator(),
+      ]);
 }