+++ lib/src/subtypes/races/gnome/gnome.dart
@@ -12,9 +12,11 @@ import '../../../generators/npcs/physical/eyes/base_eyes_generator.dart';
 import '../../../generators/npcs/physical/face/base_face_generator.dart';
 import '../../../generators/npcs/physical/hair/base_hair_generator.dart';
 import '../../../generators/npcs/physical/special_features/base_special_feature_generator.dart';
+import '../../../generators/world/opinions/base_opinion_generator.dart';
 import '../race.dart';
 import 'alignment/gnome_alignment_generator.dart';
 import 'names/gnome_name_generator.dart';
+import 'opinions/gnome_opinion_generator.dart';
 import 'quirks/gnome_quirks_generator.dart';
 import 'skin/gnome_skin_generator.dart';
 
@@ -81,4 +83,10 @@ class Gnome implements Race {
   @override
   IGenerator<String> getSpecialFeatureGenerator(Gender gender) =>
       BaseSpecialFeatureGenerator();
+
+  @override
+  IGenerator<String> getOpinionGenerator() => MultipleGenerator([
+        BaseOpinionGenerator(),
+        GnomeOpinionGenerator(),
+      ]);
 }