+++ lib/src/subtypes/races/halfling/halfling.dart
@@ -13,9 +13,11 @@ import '../../../generators/npcs/physical/face/base_face_generator.dart';
 import '../../../generators/npcs/physical/hair/base_hair_generator.dart';
 import '../../../generators/npcs/physical/skin/base_skin_generator.dart';
 import '../../../generators/npcs/physical/special_features/base_special_feature_generator.dart';
+import '../../../generators/world/opinions/base_opinion_generator.dart';
 import '../race.dart';
 import 'alignment/halfling_alignment_generator.dart';
 import 'names/halfling_name_generator.dart';
+import 'opinions/halfling_opinion_generator.dart';
 import 'quirks/halfling_quirks_generator.dart';
 
 /// A class that represents the halfling race
@@ -81,4 +83,10 @@ class Halfling implements Race {
   @override
   IGenerator<String> getSpecialFeatureGenerator(Gender gender) =>
       BaseSpecialFeatureGenerator();
+
+  @override
+  IGenerator<String> getOpinionGenerator() => MultipleGenerator([
+        BaseOpinionGenerator(),
+        HalflingOpinionGenerator(),
+      ]);
 }