+++ lib/src/subtypes/races/human/human.dart
@@ -2,6 +2,7 @@ import '../../../entities/npcs/alignment.dart';
 import '../../../entities/npcs/hair.dart';
 import '../../../enums/gender.dart';
 import '../../../generators/base/generator.dart';
+import '../../../generators/base/multiple_generator.dart';
 import '../../../generators/base/number_generator.dart';
 import '../../../generators/npcs/personality/alignment/base_alignment_generator.dart';
 import '../../../generators/npcs/personality/quirks/base_personality_quirk_generator.dart';
@@ -13,8 +14,10 @@ import '../../../generators/npcs/physical/face/base_face_generator.dart';
 import '../../../generators/npcs/physical/hair/base_hair_generator.dart';
 import '../../../generators/npcs/physical/skin/base_skin_generator.dart';
 import '../../../generators/npcs/physical/special_features/base_special_feature_generator.dart';
+import '../../../generators/world/opinions/base_opinion_generator.dart';
 import '../race.dart';
 import 'names/human_name_generator.dart';
+import 'opinions/human_opinion_generator.dart';
 
 /// A class that represents the human race
 class Human implements Race {
@@ -77,4 +80,10 @@ class Human implements Race {
   @override
   IGenerator<String> getSpecialFeatureGenerator(Gender gender) =>
       BaseSpecialFeatureGenerator();
+
+  @override
+  IGenerator<String> getOpinionGenerator() => MultipleGenerator([
+        BaseOpinionGenerator(),
+        HumanOpinionGenerator(),
+      ]);
 }