+++ lib/src/subtypes/races/tiefling/tiefling.dart
@@ -2,6 +2,7 @@ import '../../../entities/npcs/alignment.dart';
 import '../../../entities/npcs/hair.dart';
 import '../../../enums/gender.dart';
 import '../../../generators/base/generator.dart';
+import '../../../generators/base/multiple_generator.dart';
 import '../../../generators/base/number_generator.dart';
 import '../../../generators/npcs/personality/quirks/base_personality_quirk_generator.dart';
 import '../../../generators/npcs/personality/traits/personality_trait_generator.dart';
@@ -10,10 +11,12 @@ import '../../../generators/npcs/physical/build/light_build_generator.dart';
 import '../../../generators/npcs/physical/eyes/base_eyes_generator.dart';
 import '../../../generators/npcs/physical/face/base_face_generator.dart';
 import '../../../generators/npcs/physical/special_features/base_special_feature_generator.dart';
+import '../../../generators/world/opinions/base_opinion_generator.dart';
 import '../race.dart';
 import 'alignment/tiefling_alignment_generator.dart';
 import 'hair/tiefling_hair_generator.dart';
 import 'names/tiefling_name_generator.dart';
+import 'opinions/tiefling_opinion_generator.dart';
 import 'skin/tiefling_skin_generator.dart';
 
 class Tiefling implements Race {
@@ -77,4 +80,10 @@ class Tiefling implements Race {
   @override
   IGenerator<String> getSpecialFeatureGenerator(Gender gender) =>
       BaseSpecialFeatureGenerator();
+
+  @override
+  IGenerator<String> getOpinionGenerator() => MultipleGenerator([
+        BaseOpinionGenerator(),
+        TieflingOpinionGenerator(),
+      ]);
 }